fix: type correction for input_dim

+++ deeptab/architectures/experimental/modern_nca.py
@@ -93,7 +93,7 @@ class ModernNCA(BaseModel):
                 config=config,
             )
 
-            input_dim = np.sum([len(info) * self.hparams.d_model for info in feature_information])
+            input_dim = int(np.sum([len(info) * self.hparams.d_model for info in feature_information]))
         else:
             input_dim = get_feature_dimensions(*feature_information)
